chore server api caller for prefetch
