--- utils/api/header.ts
@@ -1,18 +1,0 @@
-'use server';
-
-import { cookies } from 'next/headers';
-
-export const getServerApiHeaders = async (): Promise<Headers> => {
-  const headers = new Headers({
-    'Content-Type': 'application/json',
-  });
-
-  const cookieStore = await cookies();
-  const accessToken = cookieStore.get('access_token')?.value;
-
-  if (accessToken) {
-    headers.set('Authorization', `Bearer ${accessToken}`);
-  }
-
-  return headers;
-};
